Semana2SI/requirements.txt
beautifulsoup4==4.11.1
cycler==0.11.0
fonttools==4.33.3
google==3.0.0
imutils==0.5.4
kiwisolver==1.4.2
matplotlib==3.5.2
numpy==1.22.3
opencv-python==4.5.5.64
packaging==21.3
Pillow==9.1.0
pyparsing==3.0.8
python-dateutil==2.8.2
six==1.16.0
soupsieve==2.3.2.post1
